pyproject.toml
[tool.poetry]
name = "trustek"
version = "0.1.0"
description = ""
authors = ["spiritEcosse <shevchenkcoigor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-ninja = "^1.1.0"
email-validator = "^2.1.0.post1"
django-ninja-jwt = "^5.2.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
